pyproject.toml
[project]
name = "cognitivesync"
version = "0.1.0"
description = "Privacy-first enterprise burnout prediction platform using Graph Attention Networks"
requires-python = ">=3.11"
dependencies = [
    # API layer
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    # Auth & security
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "cryptography>=42",
    # HTTP client
    "httpx>=0.27",
    # Data source connectors
    "msal>=1.28",
    "slack-sdk>=3.27",
    "PyGithub>=2.3",
    # Graph & ML
    "networkx>=3.3",
    "torch>=2.2",
    "torch-geometric>=2.5",
    # Scheduling
    "apscheduler>=3.10",
    "sqlalchemy-utils>=0.41",
    # Teams bot
    "botbuilder-core>=4.16",
    "botbuilder-schema>=4.16",
    "botbuilder-integration-aiohttp>=4.16",
    # Utilities
    "numpy>=1.26",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5",
    "respx>=0.21",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "types-networkx>=3.2",
    "types-python-jose>=3.3",
    "types-passlib>=1.7",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ANN"]
ignore = [
    "ANN101",  # Missing type annotation for `self`
    "ANN102",  # Missing type annotation for `cls`
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN"]
"data/**" = ["ANN"]

[tool.mypy]
strict = true
python_version = "3.11"
ignore_missing_imports = true
exclude = ["alembic/", "data/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=. --cov-report=term-missing --cov-report=html --cov-fail-under=80"

[tool.coverage.run]
omit = ["tests/*", "data/*", "alembic/*", "deploy/*"]

# NOTE: PyTorch with CUDA support requires configuring the PyTorch index.
# For CUDA 12.1 (RTX 4060), add the following to uv config or run:
#   uv pip install torch --index-url https://download.pytorch.org/whl/cu121
# Or add to pyproject.toml:
# [tool.uv.sources]
# torch = [{ index = "pytorch-cu121" }]
# [[tool.uv.index]]
# name = "pytorch-cu121"
# url = "https://download.pytorch.org/whl/cu121"
# explicit = true
